+++ lib/src/decorators/serialize.ts
@@ -1,48 +1,61 @@
-import Metadata from '../serializers/metadata';
-import PropertySerializer from '../serializers/property_serializer';
+import JsonPropertySerializer from '../serializers/json/json_property_serializer';
+import JsonTypeSerializer from '../serializers/json/json_type_serializer';
+import CustomTypeMetadata from '../serializers/metadata/custom_type_metadata';
+import PropertyBagMetadata from '../serializers/metadata/property_bag_metadata';
 import TypeSerializer from '../serializers/type_serializer';
+import Constructable from '../types/constructable';
 import JsonType from '../types/json_type';
 import Provider from '../types/provider';
 
 /** Decorator used to mark property for serialization with default serializer */
 function Serialize(
-    options?: PropertySerializer.Configurable.Options
+    options?: JsonPropertySerializer.Options
 ) {
     return (proto: Object, propertyName: string) => {
-        const typeSerializerProvider = () => TypeSerializer.combine([TypeSerializer.partialFor(proto, propertyName)]);
-        const propertySerializer = new PropertySerializer.Configurable(propertyName, typeSerializerProvider, options);
-        Metadata.getOrCreateFor(proto).ownSerializers.set(propertyName, propertySerializer);
+        const compiledTypeSerializerProvider = () => {
+            try {
+                return TypeSerializer.compile([JsonTypeSerializer.pickForProp(proto, propertyName)]);
+            } catch (error) {
+                const className = proto.constructor.name;
+                throw new Error(`Unable to construct a type serializer for property "${className}.${propertyName}": ${error.message}`);
+            }
+        };
+        const propertySerializer = new JsonPropertySerializer(propertyName, compiledTypeSerializerProvider, options);
+        PropertyBagMetadata.getOrCreateFor(proto).setPropertySerializer(propertyName, propertySerializer);
     };
 }
 
 namespace Serialize {
 
     /** Decorator used to mark property for serialization with custom serializer */
-    export function Custom<TSerialized extends JsonType, TOriginal = any>(
+    export function Custom<TSerialized extends JsonType, TOriginal>(
         customTypeSerializer: TypeSerializer<TSerialized, TOriginal> | Provider<TypeSerializer<TSerialized, TOriginal>>,
-        options?: PropertySerializer.Configurable.Options
+        options?: JsonPropertySerializer.Options
     ) {
         return (proto: Object, propertyName: string) => {
-            const defaultTypeSerializerProvider = () => TypeSerializer.partialFor(proto, propertyName);
+            const defaultTypeSerializerProvider = () => JsonTypeSerializer.pickForProp(proto, propertyName);
             const customTypeSerializerProvider = typeof(customTypeSerializer) === 'function' ? customTypeSerializer : () => customTypeSerializer;
-            const combinedTypeSerializerProvider = () => {
+            const compiledTypeSerializerProvider = () => {
                 try {
-                    return TypeSerializer.combine([defaultTypeSerializerProvider(), customTypeSerializerProvider()]);
+                    return TypeSerializer.compile([defaultTypeSerializerProvider(), customTypeSerializerProvider()]);
                 } catch (error) {
                     const className = proto.constructor.name;
-                    throw new Error(`Unable to construct a type serializer for "${className}.${propertyName}": ${error.message}`);
+                    throw new Error(`Unable to construct a type serializer for property "${className}.${propertyName}": ${error.message}`);
                 }
             };
-            const propertySerializer = new PropertySerializer.Configurable(propertyName, combinedTypeSerializerProvider, options);
-            Metadata.getOrCreateFor(proto).ownSerializers.set(propertyName, propertySerializer);
+            const propertySerializer = new JsonPropertySerializer(propertyName, compiledTypeSerializerProvider, options);
+            PropertyBagMetadata.getOrCreateFor(proto).setPropertySerializer(propertyName, propertySerializer);
         };
     }
 
-    /** Decorator used to explicitely mark property as non-serializable. Can be used to erase inherited serializer. */
-    export function Skip() {
-        return (proto: Object, propertyName: string) => {
-            const propertySerializer = new PropertySerializer.Dummy();
-            Metadata.getOrCreateFor(proto).ownSerializers.set(propertyName, propertySerializer);
+    /** Decorator used to define a custom serializer for a given type. */
+    export function Type<TSerialized extends JsonType, TOriginal>(
+        customTypeSerializer: TypeSerializer<TSerialized, TOriginal> | Provider<TypeSerializer<TSerialized, TOriginal>>,
+    ) {
+        return (ctor: Constructable.Default<TOriginal>) => {
+            const customTypeSerializerProvider = typeof(customTypeSerializer) === 'function' ? customTypeSerializer : () => customTypeSerializer;
+            const proto = ctor.prototype;
+            CustomTypeMetadata.getOrCreateFor(proto).setTypeSerializer(customTypeSerializerProvider);
         };
     }