+++ lib/src/serialazy.ts
@@ -1,135 +1,51 @@
 import 'reflect-metadata';
 
-import Metadata from './serializers/metadata';
+import JsonTypeSerializer from './serializers/json/json_type_serializer';
 import Constructable from './types/constructable';
-import { JsonMap } from './types/json_type';
+import { JsonType } from './types/json_type';
 
 /**
- * Deflate class instance to a JSON-compatible object
- * @param serializable Serializable class instance
- * @returns JSON-compatible object which can be safely passed to `JSON.serialize`
+ * Deflate a serializable type to a JSON-compatible type
+ * @param serializable Serializable type
+ * @returns JSON-compatible type which can be safely passed to `JSON.serialize`
  */
-export function deflate(serializable: any): JsonMap {
+export function deflate(serializable: any): JsonType {
 
-    let serialized: JsonMap;
+    let serialized: JsonType;
 
     if (serializable === null || serializable === undefined) {
-
         serialized = serializable;
-
     } else {
-
-        assertSerializable(serializable);
-        const meta = Metadata.getOwnOrInheritedMetaFor(Object.getPrototypeOf(serializable));
-
-        serialized = {};
-
-        try {
-            meta.aggregateSerializers().forEach(serializer => serializer.down(serializable, serialized));
-        } catch (error) {
-            throw new Error(`Unable to serialize an instance of a class "${meta.className}": ${error.message}`);
+        const { down } = JsonTypeSerializer.pickForValue(serializable);
+        if (!down) {
+            throw new Error(`Value is not serializable: ${serializable}`);
         }
-
+        serialized = down(serializable);
     }
 
     return serialized;
+
 }
 
 /**
- * Construct/inflate class instance from JSON-compatible object
- * @param ctor Serializable class constructor function
+ * Construct/inflate a serializable type from JSON-compatible object
+ * @param ctor Serializable type constructor function
  * @param serialized JSON-compatible object (e.g. returned from `JSON.parse`)
- * @returns Serializable class instance
+ * @returns Serializable instance
  */
-export function inflate<T>(ctor: Constructable<T>, serialized: JsonMap): T {
+export function inflate<T>(ctor: Constructable.Default<T>, serialized: JsonType): T {
 
-    let classInstance: T;
-
-    if (serialized === null || serialized === undefined) {
-
-        classInstance = serialized as any;
-
-    } else {
-
-        if (!ctor || !ctor.prototype) {
-            throw new Error('Expecting a valid constructor function');
-        }
-
-        assertSerializable(ctor);
-        const meta = Metadata.getOwnOrInheritedMetaFor(ctor.prototype);
-
-        classInstance = new ctor();
-
-        try {
-            meta.aggregateSerializers().forEach(serializer => serializer.up(classInstance, serialized));
-        } catch (error) {
-            throw new Error(`Unable to deserialize an instance of a class "${meta.className}": ${error.message}`);
-        }
-
+    if (typeof(ctor) !== 'function') {
+        throw new Error('Expecting a constructor function');
     }
 
-    return classInstance;
-}
+    const { up } = JsonTypeSerializer.pickForType(ctor);
 
-/**
- * Check if target is a serializable class constructor or an instance of serializable class
- * @param target Target to check
- */
-export function isSerializable(target: any): boolean {
-
-    if (target === null || target === undefined) {
-        throw new Error('Expecting `target` to be not null/undefined');
+    if (!up) {
+        throw new Error(`Type is not serializable: ${ctor.name}`);
     }
 
-    const meta = (typeof(target) === 'function' && target.prototype) ?
-        Metadata.getOwnOrInheritedMetaFor(target.prototype) // treat target as a constructor function
-        : Metadata.getOwnOrInheritedMetaFor(Object.getPrototypeOf(target)); // treat target as an instance
-
-    return !!meta; // treat target as serializable if it has own or inherited metadata
-
-}
-
-/**
- * Asserts that target is a serializable class constructor or an instance of serializable class
- * @param target Target to check
- */
-export function assertSerializable(target: any): void {
-
-    if (!isSerializable(target)) {
-        throw new Error(
-            `Provided instance or constructor function doesn\'t seem to be serializable: "${target}". ` +
-            'Hint: use "Serialize" decorator to mark properties for serialization'
-        );
-    }
-
-}
-
-/**
- * Traverse recursively all serializable properties of `destination`, merge them
- * with corresponding properties of `source` and return resulting object.
- * NOTE: This function mutates `destination`
- * @param destination Destination serializable class instance
- * @param source Source class instance or plain object (may be non-serializable) to take property values from
- * @returns Destination class instance
- */
-export function deepMerge<T>(destination: T, source: T): T {
-
-    if (destination === null || destination === undefined) {
-        throw new Error('Expecting `destination` to be not null/undefined');
-    }
-
-    assertSerializable(destination);
-    const meta = Metadata.getOwnOrInheritedMetaFor(Object.getPrototypeOf(destination));
-
-    if (source !== null && source !== undefined) {
-        try {
-            meta.aggregateSerializers().forEach(serializer => serializer.assign(destination, source));
-        } catch (error) {
-            throw new Error(`Unable to perform a deep property merge for instance of "${meta.className}": ${error.message}`);
-        }
-    }
-
-    return destination;
+    return up(serialized);
 
 }