+++ lib/src/serializers/type_serializer.ts
@@ -1,7 +1,6 @@
 import Constructable from '../types/constructable';
-import JsonTypeSerializer from './json/json_type_serializer';
 
-/** Represents a type serializer */
+/** Represents a generic type serializer */
 interface TypeSerializer<TSerialized, TOriginal> {
 
     /**
@@ -19,7 +18,7 @@ interface TypeSerializer<TSerialized, TOriginal> {
     up(this: void, serializedValue: TSerialized): TOriginal;
 
     /**
-     * _Optional._ Property type constructor function.
+     * _Optional._ Original type constructor function.
      * Default: Value of `design:type` for given property.
      */
     type?: Constructable.Default<TOriginal>;
@@ -27,34 +26,28 @@ interface TypeSerializer<TSerialized, TOriginal> {
     /**
      * _Optional._ Property type descriminator function.
      * Used to narrow type constructor function (e.g. for union types)
-     * @param value Property value
-     * @returns Property type constructor function
+     * @param serialized Serialized value
+     * @returns Original type constructor function
      */
-    discriminate?(this: void, value: TOriginal): Constructable.Default<TOriginal>;
+    discriminate?(this: void, serialized: TSerialized): Constructable.Default<TOriginal>;
 
 }
 
 namespace TypeSerializer {
 
-    /** Try to pick a (possibly partial) type serializer for given property */
-    export function partialFor(proto: Object, propertyName: string): Partial<JsonTypeSerializer<any>> {
-        return JsonTypeSerializer.partialFor(proto, propertyName);
-    }
-
-    /** Combine type serializer partials to final type serializer */
-    export function combine<TSerialized, TOriginal>(
+    /** Compile type serializer partials to final type serializer */
+    export function compile<TSerialized, TOriginal>(
         partials: Array<Partial<TypeSerializer<TSerialized, TOriginal>>>
     ): TypeSerializer<TSerialized, TOriginal> {
 
-        const { down, up, type, discriminate } = partials.reduce((combined, partial) => ({ ...combined, ...partial }), {});
+        const { down, up, type, discriminate } = partials.reduce((compiled, partial) => ({ ...compiled, ...partial }), {});
 
         if (!down || !up) {
             const typeName = type && type.name ? type.name : '<unknown>';
             throw new Error(
-                `No serializer for type "${typeName}". ` +
-                'Hint: Use serializable type or provide a custom serializer. ' +
-                'Try to specify property type explicitely, default serializer may be failing to pick it because of: ' +
-                'https://github.com/Microsoft/TypeScript/issues/18995. '
+                `No serializer found for type: "${typeName}". ` +
+                'Hints: Use serializable type or provide a custom serializer. ' +
+                'Specify property type explicitely, (details: https://github.com/Microsoft/TypeScript/issues/18995)'
             );
         }